Putting everything into one file, perhaps modules now found.
