--- xpsi/tests/Modules/CustomSignal.py
@@ -1,69 +1,0 @@
-
-
-import numpy as np
-import math
-
-import xpsi
-import six as _six
-
-from xpsi.likelihoods.default_background_marginalisation import eval_marginal_likelihood
-from xpsi.likelihoods.default_background_marginalisation import precomputation
-
-class CustomSignal(xpsi.Signal):
-    """ A custom calculation of the logarithm of the NICER likelihood.
-
-    We extend the :class:`xpsi.Signal.Signal` class to make it callable.
-
-    We overwrite the body of the __call__ method. The docstring for the
-    abstract method is copied.
-
-    """
-
-    def __init__(self, workspace_intervals = 1000, epsabs = 0, epsrel = 1.0e-8,
-                 epsilon = 1.0e-3, sigmas = 10.0, support = None, *args, **kwargs):
-        """ Perform precomputation. """
-
-        super(CustomSignal, self).__init__(*args, **kwargs)
-
-        try:
-            self._precomp = precomputation(self._data.counts.astype(np.int32))
-        except AttributeError:
-            print('No data... can synthesise data but cannot evaluate a '
-                  'likelihood function.')
-        else:
-            self._workspace_intervals = workspace_intervals
-            self._epsabs = epsabs
-            self._epsrel = epsrel
-            self._epsilon = epsilon
-            self._sigmas = sigmas
-
-            if support is not None:
-                self._support = support
-            else:
-                self._support = -1.0 * np.ones((self._data.counts.shape[0],2))
-                self._support[:,0] = 0.0
-
-    @property
-    def support(self):
-        return self._support
-
-    @support.setter
-    def support(self, obj):
-        self._support = obj
-
-    def __call__(self, *args, **kwargs):
-        self.loglikelihood, self.expected_counts, self.background_signal,self.background_given_support = \
-                eval_marginal_likelihood(self._data.exposure_time,
-                                          self._data.phases,
-                                          self._data.counts,
-                                          self._signals,
-                                          self._phases,
-                                          self._shifts,
-                                          self._precomp,
-                                          self._support,
-                                          self._workspace_intervals,
-                                          self._epsabs,
-                                          self._epsrel,
-                                          self._epsilon,
-                                          self._sigmas,
-                                          kwargs.get('llzero'))
